requirements.txt
streamlit==1.40.0
langchain==0.3.25
langchain-google-genai==2.0.7
langchain-community==0.3.21
faiss-cpu==1.9.0.post1
google-generativeai==0.8.0
beautifulsoup4==4.12.3
requests==2.31.0
tiktoken==0.8.0
python-dotenv==1.0.1
lxml==5.3.0